requirements.txt
# numpy==1.16.5
numpy==1.17.0
pandas==0.25.3 
scipy
scikit-learn
seaborn
jupyter 
ipykernel
matplotlib
protobuf==3.20.0
# jax[cuda12]
# jaxlib
tensorflow==1.15.0
keras==2.2.4
fair-esm
biotite
tqdm
